Creating new translation_function file which contains functions used in the translators as well as app.py.  Adding new entity api cache to reduce redundant calling of the entity api. Updating the hubmap translator to fit the translatorinterface design
